NEST/microcircuit_nest_task/PyNN_only_NEST_v1/task_requirements.txt
scipy==0.13.3
numpy==1.8.0
matplotlib==1.3.1
h5py==2.2.1
quantities==0.10.1
neo==0.4.0
elephant==0.1
tables==3.1.1
numexpr==2.4
Cython==0.21.2
PyYAML==3.10
